requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
Deprecated==1.2.13
idna==3.3
numpy==1.21.4
pandas==1.3.4
pycparser==2.20
PyGithub==1.55
PyJWT==2.3.0
pymongo==3.12.1
PyNaCl==1.4.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
soupsieve==2.3
urllib3==1.26.7
wrapt==1.13.3
